refactor: extract rise-backend-traefik, owned by Docker and ECS
The previous commit put Traefik's labelling machinery in rise-backend-core.
That was the wrong home: core is documented as the contract seam every backend
shares, and Traefik is a routing choice only two of them make. Kubernetes
routes with nginx ingress annotations and never touches any of it — a grep
across the K8s side finds zero references to labels, traefik_render,
traefik_api, diff, rolling or pod_status — so core was making every backend
depend on one backend-group's proxy.

Traefik now has its own crate, depended on by rise-backend-docker and
rise-backend-ecs:

    rise-backend-core          ← contract seam (all backends)
       └── rise-backend-traefik    ← routing (Traefik-fronted backends only)
              ├── rise-backend-docker
              └── rise-backend-ecs

The line between the two crates is principled rather than incidental: core
holds runtime-agnostic *algorithms* (env merging and hashing, drift diffing,
the rollout throttle, pod-status shaping, workload naming), the Traefik crate
holds machinery tied to one specific *proxy*. An ALB-native ECS mode would
depend on core and skip the Traefik crate entirely — which is exactly the
property that was missing before.

labels.rs split along the seam it always had: the bookkeeping vocabulary
(rise.dev/project and friends) is genuinely backend-agnostic — Docker stamps it
as container labels, ECS as resource tags, and the keys mean the same thing on
both — so it stays in core. The Traefik dynamic-config labels move. Two things
travelled with them that had been left behind in core's desired.rs:
RouteForwardAuth and the health-check label defaults, both of which are Traefik
concepts rather than desired-state ones.

The router/service naming moves too, and that one matters: group_service_base's
collision resistance is a multi-tenant security property, not a cosmetic one —
sanitize_router_name is lossy, so without its structural hash suffix two
projects could collapse onto one Traefik service and one project's labels could
suppress the other's forwardAuth. It belongs next to the labels it names.

Behaviour-preserving, and the test arithmetic shows it: core's 134 became
89 + 45 with none lost, and the workspace still passes 1227 against a real
Postgres.
