+++ CLAUDE.md
@@ -74,9 +74,10 @@ Focused support crates live under `crates/`:
 |---|---|---|
 | `rise-deployment-spec` | Shared deployment/project-config model — `AccessRequirement`, `RouteAccess`, request spec, validation | workspace dep (both CLI and backend) |
 | `rise-backend-auth` | Pure-core token signing, verification, and claim matching — the single home for auth-token logic (see `ROADMAP.md` § 2, "Rise-issued authentication and token issuance") | `backend` |
-| `rise-backend-core` | The deployment-backend contract seam: shared deployment models, the `DeploymentBackend` trait, registry/encryption provider traits, the pure `quantity`/`state_machine`/`runtime`/`url_builder`/`token_ttl`/`custom_domain` helpers, the `DeploymentStore` trait — the database boundary implemented by `rise-deploy`'s `PgDeploymentStore` — and the runtime-agnostic reconcile machinery every backend shares (`desired`/`env`/`naming`/`labels`/`traefik_render`/`traefik_api`/`diff`/`rolling`/`pod_status`) | `backend` |
+| `rise-backend-core` | The deployment-backend contract seam: shared deployment models, the `DeploymentBackend` trait, registry/encryption provider traits, the pure `quantity`/`state_machine`/`runtime`/`url_builder`/`token_ttl`/`custom_domain` helpers, the `DeploymentStore` trait — the database boundary implemented by `rise-deploy`'s `PgDeploymentStore` — and the runtime-agnostic reconcile machinery every backend shares (`desired`/`env`/`naming`/`labels`/`diff`/`rolling`/`pod_status`). Deliberately free of any one proxy's machinery: routing lives in `rise-backend-traefik` | `backend` |
 | `rise-backend-docker` | The Docker deployment backend: `DockerBackend` + the in-process `DockerReconciler`, the first controller extracted onto the `rise-backend-core` seam. Re-exported under `crate::server::deployment::controller::docker`, so existing module paths keep resolving | `backend` |
 | `rise-backend-ecs` | The Amazon ECS deployment backend (Fargate): `EcsBackend` + the in-process `EcsReconciler`, plus the pure Fargate-sizing, tag, task-definition and service-diff modules | `backend` |
+| `rise-backend-traefik` | Traefik routing machinery — the dynamic-config label vocabulary, the `DesiredContainer` → labels renderer with its fail-closed `routes_withheld` predicate, router/service naming, the `serverStatus` API client, and the readiness verdicts built on it. Owned by the Traefik-fronted backends (Docker, ECS) rather than by `rise-backend-core`: Kubernetes routes with nginx annotations and depends on none of it | `backend` |
 | `rise-authz` | Authorization policy evaluation, in two tiers. `policy` is a hard Tier-0 boundary: pure functions and canonical values only — no store, database, HTTP, or product-resource dependencies. `engine` is Tier 1: the live ADR-0001 §4 algorithm over `ResourceStore` reads and one `MembershipResolver` seam — still free of HTTP, SQL, and product kinds | `backend` |
 | `rise-resource-api` | Generic resource API contract (resource kinds, scopes, owner references, identity, policy types) | `backend` |
 | `rise-resource-store-postgres` | PostgreSQL adapter for the resource API. Owns its own migrations in a `resource_store` schema and its own SQLX offline cache | `backend` |
