+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "crates/rise-backend-core",
     "crates/rise-backend-docker",
     "crates/rise-backend-ecs",
+    "crates/rise-backend-traefik",
     "crates/rise-deployment-spec",
     "crates/rise-resource-api",
     "crates/rise-resource-store-postgres",
@@ -87,6 +88,7 @@ backend = [
     "dep:rise-backend-core",
     "dep:rise-backend-docker",
     "dep:rise-backend-ecs",
+    "dep:rise-backend-traefik",
     "dep:rise-resource-api",
     "dep:rise-resource-store-postgres",
     "dep:rise-runtime-sync",
@@ -140,6 +142,7 @@ rise-backend-auth = { path = "crates/rise-backend-auth", optional = true }
 rise-backend-core = { path = "crates/rise-backend-core", optional = true }
 rise-backend-docker = { path = "crates/rise-backend-docker", optional = true }
 rise-backend-ecs = { path = "crates/rise-backend-ecs", optional = true }
+rise-backend-traefik = { path = "crates/rise-backend-traefik", optional = true }
 rise-resource-api = { path = "crates/rise-resource-api", optional = true }
 rise-resource-store-postgres = { path = "crates/rise-resource-store-postgres", optional = true }
 rise-runtime-sync = { path = "crates/rise-runtime-sync", optional = true }
