+++ crates/rise-backend-docker/Cargo.toml
@@ -16,6 +16,7 @@ futures = "0.3"
 reqwest = { version = "0.13.0", features = ["json", "stream", "form"] }
 rise-backend-auth = { path = "../rise-backend-auth" }
 rise-backend-core = { path = "../rise-backend-core" }
+rise-backend-traefik = { path = "../rise-backend-traefik" }
 rise-deployment-spec = { path = "../rise-deployment-spec" }
 rise-resource-api = { path = "../rise-resource-api" }
 rise-runtime-sync = { path = "../rise-runtime-sync" }
